HealthTrace/data-pipeline/requirements.txt
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
kafka-python==2.0.2
requests==2.31.0
aiohttp==3.9.1
asyncio
